Update reset api
Moved the 8 bit specific API entry point up to the base class. Did this to allow dropping 16 bit chips into code written for the 8 bit chips without having to recompile. Note that the GPIO style members on the API are agnostic and simply check to see if the pin numbers requested are in range.

Factored the reset api the same way as the write/read api (ie port specificaion is only available on 16 bit chips).

Finished updating the sample program. Removed one sample that we don't support directly (raw writes to register numbers). This is supported through the lower level bus drivers or via inheritance (proitected members) if really needed.