Don't fail the build when myget publishing fails
There is an ongoing issue with myget which has caused that most of our builds in dotnet/iot have been red for some time. Given that myget is not our main publishing feed anymore, we don't want the builds to fail while we wait for the fix in the feed, so setting continueOnError in order to only throw a warning.