Fast interrupt handling (#914)
* Special overload for the SysFsDriver when used as child of the Raspi driver

Removed unneccessary delays and status handling for
pins inside the SysFsDriver when the Raspi driver
is the main driver. This can now handle interrupts
with frequencies of several khz.

Includes a test, but this is not reliable

* Check whether the test fails due to missing thread priority

* Code cleanup

* It seems to work without the extra priority now

I don't really understand why it seems to be all fine now, though.

* This part is unnecessary

That's the documented way of reading the new value,
but we're not using the result here and open the value file
later again in the Read() function.

* Avoid the same error text in multiple places

* Read out actual new value of pin after interrupt

This fixes #925

* Add separate member for actual event observed

Note that the Windows implementation already returned the actual event type,
while the Linux implementation returned the desired event type(s).

* Remove Pin status validation from Driver class, done by controller

Keep track of which pins are in use by SysFs

* Remove unused method

PinMode is never set trough SysFs driver any more

* Return detected event correctly in all cases (#925)

... but use the existing member, as discussed.

* Avoid issues with missing permissions

* Keep track of which events are active

Like this, if the edge mode is not both, we do know
for sure which one we saw

* Fix typo

* This test fails sometimes.

Maybe due to an event on another line?

* The AddCallbackFallingEdgeNotDetectedTest requires some extra code to be reliable

Sometimes, an additional event is triggered when registering the event,
causing the test to fail (because an event is seen where none is expected)

* Cleanup

* Ensure LibGpiodDriver is thread safe

And allow repeatedly opening the same pin

* This test may get an additional event as well

* This test should now work against all drivers

It only seems to fail on Windows, reason currently unknown

* Review findings

- Remove useless sysFsInitializationLock
- Replace Dictionary with Array (size is known)
- Rename "SysFs" with "Interrupt", as the driver can now
  also be libgpiod

* Check why Windows test is failing (temporary)

* Make Windows test pass (eventually?)

* Don't know why this takes longer

* Remove lock for Read/Write/SetPinMode operations

Can be done when using ConcurrentDictionary

* Use explicit type

* Don't use static ctor

* Remove unnecessary property, change type

* Read pin state immediately

Right after an eventual wait, if required.

* See comment and bug #1028