Interop.libgpiod: Add gpiod_line_find
From Linux Kernel docs[1]:
"Optionally, a GPIO controller may have a "gpio-line-names" property.
This is an array of strings defining the names of the GPIO lines going
out of the GPIO controller. This name should be the most meaningful
producer name for the system, such as a rail name indicating the usage."

gpiod_line_find find for a gpio line that was defined in
gpio-line-names and return its handle. This is very useful to avoid
having to search in the datasheets the pin bank/line. The user can
check the labels with the pin names on the PCB and use it instead.

1: https://www.kernel.org/doc/Documentation/devicetree/bindings/gpio/gpio.txt

Signed-off-by: Matheus Castello <matheus@castello.eng.br>