Make all internal types in Common internal OR public (#1759)
This cleans up a few dependency issues between the bindings and the common project. The classes marked internal (including the Interop classes) are now public when building Common.csproj but still internal for the final assembly. This removes the need to link to individual source files outside their binding folders, which causes build problems when they're used from multiple places.