MFRC522 tranceiver support for Write16Bytes
Mifare WRite16Bytes, which is used by Mifare Classic and Mifare Ultralight (where it is called
"compatibility write") requires two writes to the card. The first one transfers the
block number, and the second contains the data to be written. Changed the Transceive
function to handle Write16Bytes the same as counter increment, decrement, and restore
(which also require two writes to the card).