Added CM4 model check to IsPi4
CM4 needs to be handled the same way as a pi4. Right now, due to the device model of a CM4 being "Raspberry Pi Computer Module 4" it is being treated as a pi3. This causes a problem with the pin mapping and prevents the CM4 from having it's input/output and pullup/pulldown changed on OpenPin().