Give better wiring instructions for 1-wire devices.
The provided diagram is wrong (10k resistor is in series with the data line, leading to unreliable operation).  I believe it was intended as a pullup (as recommended by the datasheet)

However, since the pi IO pins are not 5v tolerant, to do that we need to pull up to 3.3v.

To avoid confusion, we also power the sensor from 3.3v (the datasheet allows this).